requirements.txt
aiohttp==3.8.5
aiosignal==1.3.1
anyio==3.7.1
async-timeout==4.0.2
asyncio==3.4.3
attrs==23.1.0
autopep8==2.0.2
beautifultable==1.1.0
certifi==2023.7.22
charset-normalizer==3.2.0
click==8.1.6
colorama==0.4.6
copier==8.1.0
decorator==5.1.1
dill==0.3.7
dunamai==1.18.0
exceptiongroup==1.1.2
filelock==3.12.2
flet==0.9.0
flet-core==0.9.0
flet-runtime==0.9.0
frozenlist==1.4.0
funcy==2.0
h11==0.14.0
httpcore==0.17.3
httpx==0.24.1
idna==3.4
importlib-metadata==6.8.0
Jinja2==3.1.2
jinja2-ansible-filters==1.3.2
linkify-it-py==2.0.2
markdown-it-py==3.0.0
MarkupSafe==2.1.3
mdit-py-plugins==0.4.0
mdurl==0.1.2
mopyx==1.1.1
msgpack==1.0.5
multidict==6.0.4
oauthlib==3.2.2
packaging==23.1
pathspec==0.11.2
plumbum==1.8.2
pprintpp==0.4.0
prompt-toolkit==3.0.36
pycodestyle==2.10.0
pydantic==1.10.12
pydux==0.2.2
pyfunctional==1.4.3
Pygments==2.15.1
pypng==0.20220715.0
pysondb==1.6.7
pywin32==306
PyYAML==6.0.1
pyyaml-include==1.3.1
qrcode==7.4.2
questionary==2.0.0
repath==0.9.0
requests==2.31.0
rich==13.5.2
six==1.16.0
sniffio==1.3.0
tabulate==0.9.0
textual==0.32.0
textual-dev==1.1.0
tomli==2.0.1
typing_extensions==4.7.1
uc-micro-py==1.0.2
Unidecode==1.3.6
urllib3==2.0.4
watchdog==3.0.0
wcwidth==0.2.6
websocket-client==1.6.1
websockets==11.0.3
yarl==1.9.2
zipp==3.16.2
